refactor structure of UI, add devextreme
