+++ restaurant.client/index.html
@@ -7,7 +7,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Vite + React</title>
   </head>
-  <body>
+  <body class="dx-viewport">
     <div id="root"></div>
     <script type="module" src="/src/main.jsx"></script>
   </body>
