+++ restaurant.client/package.json
@@ -13,6 +13,8 @@
     "@remixicon/react": "^4.1.1",
     "axios": "^1.6.7",
     "bootstrap": "^5.3.2",
+    "devextreme": "23.1.4",
+    "devextreme-react": "23.1.4",
     "formik": "^2.4.5",
     "react": "^18.2.0",
     "react-bootstrap": "^2.10.0",
