+++ restaurant.client/src/App.jsx
@@ -1,12 +1,13 @@
 import "./App.css";
 import { RouterProvider } from "react-router";
-import Router from "./Router";
-import { AuthProvider } from "./User/AuthContext";
+import UserRouter from "./Routes/UserRouter";
+import { AuthProvider } from "./state/Auth/AuthContext";
+import 'devextreme/dist/css/dx.material.lime.light.css';
 
 const App = () => {
   return (
     <AuthProvider>
-      <RouterProvider router={Router} />
+      <RouterProvider router={UserRouter} />
     </AuthProvider>
   );
 };
