+++ restaurant.client/src/Layout/UserLayout/Layout.jsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { Outlet } from "react-router";
 
-import { OrderProvider } from "../Cart/OrderContext.jsx";
-import Header from "./Header.jsx";
-import Footer from "./Footer.jsx";
+import { OrderProvider } from "../../state/Order/OrderContext.jsx";
+import Header from "../Header.jsx";
+import Footer from "../Footer.jsx";
 import "./Layout.css";
 
 const Layout = () => {
