+++ restaurant.client/src/components/Cart/CartItemsList.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Row } from "react-bootstrap";
 import CartItem from "./CartItem";
-import { useOrder } from "./OrderContext";
+import { useOrder } from "../../state/Order/OrderContext";
 
 const CartItemsList = () => {
   const order = useOrder();
