+++ restaurant.client/src/state/Auth/AuthContext.jsx
@@ -5,20 +5,23 @@ export const AuthContext = createContext();
 export const AuthProvider = ({ children }) => {
   const [isAuthenticated, setIsAuthenticated] = useState(false);
   const [user, setUser] = useState(null);
+  const [userRole, setUserRole] = useState(null);
 
   const login = (user) => {
     setIsAuthenticated(true);
     setUser(user);
+    setUserRole(user.role);
   };
 
   const logout = () => {
     setIsAuthenticated(false);
     setUser(null);
+    setUserRole(null);
   };
 
   return (
     <AuthContext.Provider
-      value={{ isAuthenticated, login: login, logout: logout, user }}
+      value={{ isAuthenticated, login: login, logout: logout, user, userRole}}
     >
       {children}
     </AuthContext.Provider>
